Code review comments refactor.
