+++ packages/babel-cli/src/babel/index.js
@@ -4,11 +4,11 @@
 require("babel-core");
 
 let fs         = require("fs");
-let child      = require("child_process");
 let commander  = require("commander");
 let kebabCase  = require("lodash/kebabCase");
 let options    = require("babel-core").options;
 let util       = require("babel-core").util;
+let cliutil    = require("./util");
 let uniq       = require("lodash/uniq");
 let each       = require("lodash/each");
 let glob       = require("glob");
@@ -128,51 +128,8 @@ if (commander.outDir) {
   fn = require("./file");
 }
 
-if(commander.settings) {
-  let fileOptionsManager = require("babel-core").File,
-      allOptions = [];
-  let generalOptions = new fileOptionsManager( { filename: "unknown" } ).opts;
-
-  allOptions.push(generalOptions);
-
-  each(filenames, function (file) {
-    let fileOptions = new fileOptionsManager( { filename: file } ).opts,
-        thisFileOptions = {};
-
-    each(Object.keys(fileOptions), function (key) {
-      if(!generalOptions.hasOwnProperty(key) || JSON.stringify(generalOptions[key]) != JSON.stringify(fileOptions[key])) {
-        thisFileOptions[key] = fileOptions[key];
-      }
-    });
-    allOptions.push(thisFileOptions);
-  });
-
-  let printObject = function (filename, opts) {
-    console.log(`--- ${filename} options ---`);
-    each(Object.keys(opts), function (key) {
-      console.log(key, ": ", opts[key]);
-    });
-    console.log();
-  };
-
-  console.log(`\n========== SETTINGS ==========\n`);
-  process.stdout.write(`node version: `);
-  child.execSync("node -v", {stdio:[0, 1]});
-  process.stdout.write(`npm version: `);
-  child.execSync("npm -v", {stdio:[0, 1]});
-  process.stdout.write(`packages:\n`);
-  child.execSync("npm list", {stdio:[0, 1]});
-
-  each(allOptions, function (fileOptions, index) {
-    if (index !== 0) {
-      printObject(fileOptions.filename, fileOptions);
-    } else {
-      let header = `General ${process.cwd()}`;
-      printObject(header, fileOptions);
-    }
-  });
-  console.log(`======= END OF SETTINGS ======\n`);
-  process.exit(0);
+if (commander.settings) {
+  cliutil.getSettings(filenames);
 }
 
 fn(commander, filenames, exports.opts);
